Add XHR header to search request.
The rationale behind this addition is in the comments added. Put simply, I'm making the Flask app use Flask's jsonify function in order to simplify life, but jsonify() uses indent=2 if the request isn't an XHR. From experience, the request takes much longer to process when the JSON is not in condensed form (clearly because the readLine() loop goes around many more times), so I'll add this simple change to make things work smoothly.

Also, time to go to bed... Ugh.